Update init.pp
Added $stanza variable and changed module to use that to build the changes. This solves the problem of not being able to modify the same stanza in two different conf files. This is an issue when you have Splunk and Universal Forwarder installed on the same box and want to edit the web.conf for both, for instance.